backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
asyncpg==0.29.0
pydantic==2.7.0
pydantic[email]
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
PyJWT==2.8.0
yahooquery==2.3.7
openpyxl==3.1.2
python-multipart==0.0.9
apscheduler==3.10.4
httpx==0.27.0
